🔨 refactor: my page profile edit shema 함수 분리
